+++ Model/Configuration.php
@@ -21,7 +21,7 @@ use Magento\Config\Model\ResourceModel\Config;
 
 class Configuration
 {
-    const TAXJAR_VERSION              = '1.3.0';
+    const TAXJAR_VERSION              = '1.4.2';
     const TAXJAR_AUTH_URL             = 'https://app.taxjar.com';
     const TAXJAR_API_URL              = 'https://api.taxjar.com/v2';
     const TAXJAR_FEED_URL             = 'www.taxjar.com/magento2/feed.xml';
@@ -45,6 +45,7 @@ class Configuration
     const TAXJAR_CALCULATIONS_LOG     = 'calculations.log';
     const TAXJAR_TRANSACTIONS_LOG     = 'transactions.log';
     const TAXJAR_ADDRVALIDATION_LOG   = 'address_validation.log';
+    const TAXJAR_CUSTOMER_LOG         = 'customers.log';
     const TAXJAR_EXEMPT_TAX_CODE      = '99999';
 
     /**