+++ view/adminhtml/requirejs-config.js
@@ -18,8 +18,7 @@ var config = {
     map: {
         '*': {
             taxjarClipboard: 'Taxjar_SalesTax/js/clipboard',
-            taxjarPopup: 'Taxjar_SalesTax/js/popup',
-            taxjarModal: 'Taxjar_SalesTax/js/modal'
+            taxjarPopup: 'Taxjar_SalesTax/js/popup'
         }
     }
 };